Add task-based MSC class driver
This replaces the default MSC class driver with one that pushes
the processing of the SCSI messages to a task to avoid
I/O operations from ISR context.

Also, the FSS pin specification cannot be left unset (0) as this
causes data corruption. For now, set the pin to the default.
Later, the CS functionality will be taken over via GPIO_setConfig
for the corresponding pins.