Add ctrl-c keyboard interrupt for USB UART
This adds a hook fxn to the CDCD UART receive logic so input
can be filterd for ctrl-c. This is enabled by MICROPY_KBD_EXCEPTION.

This appears to break the rshell utility (and there are some
posts about this for other ports) but mpfshell and pyboard work
with my limited testing.

Note that serial UART still works but will not respond to ctrl-c.