Performance tweaks for ILI9341 display driver
This uses direct access to the SPI registers to optimize the
one byte transfers for data and cmd. Also tweaks copy loop
for slight improvement. These make approximately 25% improvement
in a couple small synthetic benchmarks - YMMV. This optimization
is SoC-specific since most devices do not share a common
SPI controller.

Bumping the SPI clock to 30MHz helps fills/clears look much
crisper.

Noticed that the size of the DMA buffer seems to make a difference,
but not intuitively to me. I thought a size of 320 or 240 would
be best but some testing seems to indicate not.