neopix: fix neopix 'reset' period
To make the reset period stay low we need to enable the TAPLO
mode in GPTMTAMR and set the match value equal to the load value.

The TRM (SLAU723) describes how the PWM operates when PLO is set
but the code was leaving it in legacy mode. There is no obvious
API for enabling this bit but we can set it via TimerUpdateMode().